// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  Student: 'Student',
  Teacher: 'Teacher',
  Admin: 'Admin'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const AttendanceStatus = {
  NotMarked: 'NotMarked',
  Present: 'Present',
  Absent: 'Absent',
  Late: 'Late'
} as const

export type AttendanceStatus = (typeof AttendanceStatus)[keyof typeof AttendanceStatus]


export const DayType = {
  Working: 'Working',
  Holiday: 'Holiday',
  Exam: 'Exam',
  Event: 'Event',
  Weekend: 'Weekend'
} as const

export type DayType = (typeof DayType)[keyof typeof DayType]


export const Stream = {
  Science: 'Science',
  Commerce: 'Commerce'
} as const

export type Stream = (typeof Stream)[keyof typeof Stream]


export const SecondLanguage = {
  Kannada: 'Kannada',
  Hindi: 'Hindi',
  Sanskrit: 'Sanskrit'
} as const

export type SecondLanguage = (typeof SecondLanguage)[keyof typeof SecondLanguage]